requirements.txt
PyQt5
psutil
Pillow
